refactor : LottoWinningTicket -> LottoTicket 클래스로 통일
