+++ src/main/java/lotto/domain/LottoTicket.java
@@ -1,5 +1,8 @@
 package lotto.domain;
 
+import lotto.exception.NotANumberException;
+import lotto.exception.NumberDuplicateException;
+import lotto.exception.OutOfRangeException;
 import lotto.util.LottoTicketUtils;
 
 import java.util.ArrayList;
@@ -13,12 +16,17 @@ public class LottoTicket {
         this.ticket = autoNumberGenerate();
     }
 
+    public LottoTicket(List<Integer> numbers) {
+        checkLengthValidation(numbers);
+        this.ticket = checkNumberValidation(numbers);
+    }
+
     public List<Integer> showTicketNumber() {
         return ticket;
     }
 
-    public int findMatchCount(LottoWinningTicket lottoWinningTicket) {
-        List<Integer> copyList = new ArrayList<>(lottoWinningTicket.getWinningNumber());
+    public int findMatchCount(LottoTicket lottoWinningTicket) {
+        List<Integer> copyList = new ArrayList<>(lottoWinningTicket.showTicketNumber());
         copyList.removeAll(ticket);
         return LottoTicketUtils.RANGE - copyList.size();
     }
@@ -35,5 +43,33 @@ public class LottoTicket {
 
         Collections.sort(lotto);
         return lotto;
+    }
+
+    private void checkLengthValidation(List<Integer> numberInformation) {
+        if (numberInformation.size() != LottoTicketUtils.RANGE) {
+            throw new OutOfRangeException();
+        }
+    }
+
+    private List<Integer> checkNumberValidation(List<Integer> numbers) {
+        List<Integer> winningNumber = new ArrayList<>();
+        for (int number : numbers) {
+            checkNumberRange(number);
+            checkDuplicateThenAdd(winningNumber, number);
+        }
+        return winningNumber;
+    }
+
+    private void checkDuplicateThenAdd(List<Integer> winningNumber, int number) {
+        if (winningNumber.contains(number)) {
+            throw new NumberDuplicateException();
+        }
+        winningNumber.add(number);
+    }
+
+    private void checkNumberRange(int number) {
+        if (number < LottoTicketUtils.TICKET_MIN_NUMBER || number > LottoTicketUtils.TICKET_MAX_NUMBER) {
+            throw new NotANumberException(number);
+        }
     }
 }
