+++ src/test/java/lotto/LottoWinningTest.java
@@ -1,8 +1,8 @@
 package lotto;
 
 import lotto.domain.LottoMachine;
+import lotto.domain.LottoTicket;
 import lotto.domain.LottoTickets;
-import lotto.domain.LottoWinningTicket;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
@@ -15,23 +15,23 @@ public class LottoWinningTest {
     @Test
     void winningLottoMakeTest() {
          List<Integer> numbers = Arrays.asList(1,2,3,4,5,6);
-        LottoWinningTicket winningTicket = new LottoWinningTicket(numbers);
+        LottoTicket winningTicket = new LottoTicket(numbers);
         assertThat(winningTicket).isNotNull();
     }
 
     @Test
     void winningLottoExceptionTest() {
         List<Integer> numbers = Arrays.asList(1,2,3,4);
-        assertThatThrownBy(() ->new LottoWinningTicket(numbers))
+        assertThatThrownBy(() ->new LottoTicket(numbers))
                 .isInstanceOf(RuntimeException.class);
     }
 
     @Test
     void winningLottoMatchingTest() {
         List<Integer> winningNumber = Arrays.asList(1,5,6,9,10,13);
-        LottoWinningTicket lottoWinningTicket = new LottoWinningTicket(winningNumber);
+        LottoTicket LottoTicket = new LottoTicket(winningNumber);
         LottoTickets lottoTickets = new LottoTickets(100);
-        LottoMachine.winningResult(lottoWinningTicket, lottoTickets);
+        LottoMachine.winningResult(LottoTicket, lottoTickets);
     }
